ml_service/requirements.txt
fastapi
uvicorn
python-multipart
numpy
pillow
opencv-python
tensorflow